added new functions to table and plot
This commit also adds new chat actions based on the function call. Function calls for table and plot functions are considered EOS. The chat answer dictionary can contain a table or plot along with the SQL. SQL is now updated with each call to fetch_data. There is a condition to if SQL was returned to streamlit.